+++ app/src/main/java/de/kochon/enrico/secrettalkmessenger/backend/ChannelCacheRefreshable.java
@@ -2,7 +2,7 @@ package de.kochon.enrico.secrettalkmessenger.backend;
 
 public interface ChannelCacheRefreshable {
 
-   public void refreshConversationsFromCache(int idchannel);
+   public void refreshConversationsFromCache(long idchannel);
 
    public void indicateRefresh();
    public void stopRefreshIndication();
